veiculos/requirements.txt.txt
streamlit>=1.28.0
mysql-connector-python>=8.1.0
pandas>=2.1.0
plotly>=5.17.0
requests>=2.31.0
python-docx>=0.8.12
fpdf2>=2.7.5
openpyxl>=3.1.2